requirements.txt
Django==5.1
python-dotenv==1.0.0
psycopg2==2.9.9
djangorestframework==3.15.2
PyMuPDF==1.25.3